pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "gaSombrio", email = "sombrio.ga@gmail.com"},
]
dependencies = ["django>=5.2.5", "pillow>=11.3.0", "djangorestframework>=3.16.1"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
